Switch from HttpBasic to OAuth2 resource server
Transitioned the application's security configuration from using basic Http authentication to relying on an OAuth2 JWT resource server. This change was made to increase the security of the application, as well as to align with industry best practices. Removed Basic Security user credentials from application properties file and added a dependency for OAuth2 server in pom.xml. Disabled the CSRF protection for API endpoints as it's unnecessary with JWT based authentication.