Replace basic authentication with JWT in tests
In tests for the CustomerController and BeerController, the basic authentication used in requests has been replaced with JWT authentication. This change was made to reflect the shift from basic authentication to JWT in the application. Also, JWT RequestPostProcessor had been added as a constant. This JWT processor has claims for "message-read" and "message-write" scopes and uses a subject "messaging-client".