Refactor Controller methods and Maven dependencies
Refactored several methods in CustomErrorController, CustomerController, and BeerController to improve readability and efficiency. Whitespace is adjusted, ResponseEntity is returned more elegantly, and fixed the URI creation method for PostMapping.

The controller package in the test folder is updated to include @Disabled for a test method planned to be fixed later.

Updated the Pom.xml to upgrade Spring Boot version from 3.1.3 to 3.2.0 and adjusted the mysql dependency to use com.mysql instead of mysql groupID with runtime scope.

All these changes will provide a cleaner code base, expedite future developments and ensure compatibility with the latest Spring Boot version.