Simplified UUID generation strategy and skipped tests during packaging
Streamlined the way UUIDs are generated for the `Customer` and `Beer` entities. The previous method was utilizing an unnecessarily complex approach which has now been replaced with a simpler `GenerationType.UUID` for more efficient generation.

Additionally, during Maven's build process, the 'run tests' case has been skipped when packaging the application to save time and resources when the test suite is extensive. Note that tests are still executed separately in their designated step.